% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plot.R
\name{get_plot}
\alias{get_plot}
\title{Get User Plot}
\usage{
get_plot(df, playType, plot)
}
\arguments{
\item{df}{NFL data frame}

\item{playType}{Special Teams play type}

\item{plot}{Type of plot to get}
}
\value{
A ggplot object containing the user requested plot
}
\description{
Get User Plot
}
