% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_plots.R
\name{getPlot_hangTime}
\alias{getPlot_hangTime}
\title{Title}
\usage{
getPlot_hangTime(df, playType)
}
\arguments{
\item{df}{A dataframe containing information on NFL plays}

\item{playType}{One of the NFL playtypes Extra Point, Field Goal, Kickoff or Punt.}
}
\value{
A ggplot figure containing histogram of hangTime
}
\description{
Title
}
