% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_plots.R
\name{getPlot_kick_playType}
\alias{getPlot_kick_playType}
\title{Kick Length Distribution Plots}
\usage{
getPlot_kick_playType(df, playType)
}
\arguments{
\item{df}{A dataframe containing information on NFL plays}

\item{playType}{One of the NFL play types Extra Point, Field Goal, Kickoff or
Punt.}
}
\value{
A ggplot figure containing histogram of `kickLength`
}
\description{
Kick Length Distribution Plots
}
