% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryPlots.R
\name{getPlot_kick_result}
\alias{getPlot_kick_result}
\title{Title}
\usage{
getPlot_kick_result(df, playType)
}
\arguments{
\item{playType}{}
}
\value{

}
\description{
Title
}
\examples{
https://r-pkgs.org/man.html#man-functions
}
