% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_plots.R
\name{getPlot_topKickers}
\alias{getPlot_topKickers}
\title{Title}
\usage{
getPlot_topKickers(df, playType)
}
\arguments{
\item{df}{A dataframe containing information on NFL plays}

\item{playType}{One of the NFL playtypes Extra Point, Field Goal, Kickoff or Punt.}
}
\value{
A ggplot figure containing top players with accurate kicks
}
\description{
Title
}
