% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plot_list.R
\name{make_plot_list}
\alias{make_plot_list}
\title{Plot List By Playtype}
\usage{
make_plot_list(playtype = "Punt")
}
\arguments{
\item{playtype}{Special teams play type}
}
\value{
A list with plots available.
}
\description{
Plot List By Playtype
}
