% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PFFScoutingData}
\alias{PFFScoutingData}
\title{Contains data on scouting information}
\format{
A data frame with 19979 rows and 20 variables: \describe{
  \item{gameId}{Game identifier, unique} \item{playId}{Play identifier, not
  unique across games} \item{snapDetail}{On Punts, whether the snap was on
  target and if not, provides detail (H: High, L: Low, <: Left, >: Right, OK:
  Accurate Snap, text)}\item{snapTime}{Time of the snap}
  \item{operationTime}{Timing from snap to kick on punt plays in seconds: }
  \item{hangTime}{Hangtime of player's punt or kickoff attempt in seconds.
  Timing is taken from impact with foot to impact with the ground or a
  player. (numeric)} \item{kickType}{Kickoff or Punt Type. Possible values
  for kickoff plays:,D: Deep - your normal deep kick with decent hang time,F:
  Flat - different than a Squib in that it will have some hang time and no
  roll but has a lower trajectory and hang time than a Deep kick off,K: Free
  Kick - Kick after a safety,O: Obvious Onside - score and situation dictates
  the need to regain possession. Also the hands team is on for the returning
  team,P: Pooch kick - high for hangtime but not a lot of distance - usually
  targeting an upman,Q: Squib - low-line drive kick that bounces or rolls
  considerably, with virtually no hang time,S: Surprise Onside - accounting
  for score and situation an onsides kick that the returning team doesn’t
  expect. Hands teams probably aren't on the field,B: Deep Direct OOB -
  Kickoff that is aimed deep (regular kickoff) that goes OOB directly
  (doesn't bounce),Possible values for punt plays:,N: Normal - standard punt
  style,R: Rugby style punt,A: Nose down or Aussie-style punts}
  \item{kickDirectionIntended}{Intended kick direction from the kicking
  team's perspective - based on how coverage unit sets up and other factors
  (L: Left, R: Right, C: Center, text).} \item{kickDirectionActual}{Actual
  kick direction from the kicking team's perspective (L: Left, R: Right, C:
  Center, text).} \item{returnDirectionIntended}{The return direction the
  punt return or kick off return unit is set up for from the return team's
  perspective (L: Left, R: Right, C: Center, text).}
  \item{returnDirectionActual}{Actual return direction from the return team's
  perspective (L: Left, R: Right, C: Center, text).}
  \item{missedTackler}{Jersey number and team code of player(s) charged with
  a missed tackle on the play. It will be reasonable to assume that he should
  have brought down the ball carrier and failed to do so. This situation does
  not have to entail contact, but it most frequently does. Missed tackles on
  a QB by a pass rusher are also included here. Multiple missed tacklers on a
  play are separated by a } \item{assistTackler}{Jersey number and team code
  of player(s) assisting on the tackle. Multiple assist tacklers on a play
  are separated by a} \item{tackler}{Jersey number and team code of player
  making the tackle} \item{kickoffReturnFormation}{3 digit code indicating
  the number of players in the Front Wall, Mid Wall and Back Wall}
  \item{gunners}{Jersey number and team code of player(s) lined up as gunner
  on punt unit. Multiple gunners on a play are separated by a}
  \item{puntRushers}{Jersey number and team code of player(s) on the punt
  return unit with "Punt Rush" role for actively trying to block the punt.
  Does not include players crossing the line of scrimmage to engage in punt
  coverage players in a "Hold Up" role. Multiple punt rushers on a play are
  separated by a} \item{specialTeamsSafeties}{Jersey number and team code for
  player(s) with "Safety" roles on kickoff coverage and field goal/extra
  point block units - and those not actively advancing towards the line of
  scrimmage on the punt return unit. Multiple special teams safeties on a
  play are separated by a} \item{vises}{Jersey number and team code for
  player(s) with a "Vise" role on the punt return unit. Multiple vises on a
  play are separated by a} \item{kickContactType}{Detail on how a punt was
  fielded, or what happened when it wasn't fielded. Possible values: BB:
  Bounced Backwards,BC: Bobbled Catch from Air,BF: Bounced Forwards,BOG:
  Bobbled on Ground,CC: Clean Catch from Air,CFFG: Clean Field From
  Ground,DEZ: Direct to Endzone,ICC: Incidental Coverage Team Contact,KTB:
  Kick Team Knocked Back,KTC: Kick Team atch,KTF: Kick Team Knocked
  Forward,MBC: Muffed by Contact with Non-Designated Returner,MBDR: Muffed by
  Designated Returner,OOB: Directly Out Of Bounds} }
}
\source{
\url{https://www.kaggle.com/c/nfl-big-data-bowl-2022/data}
}
\usage{
PFFScoutingData
}
\description{
A dataset containing play-level scouting information for each game.
}
\keyword{datasets}
