% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PLAYERS}
\alias{PLAYERS}
\title{Contains data on players.}
\format{
A data frame with 2732 rows and 7 variables: \describe{
  \item{nflId}{Player identification number, unique across players}
  \item{height}{Player height} \item{weight}{Player weight}
  \item{birthDate}{Date of birth (YYYY-MM-DD)} \item{collegeName}{Player
  college} \item{Position}{Player position} \item{displayName}{Player
  name} }
}
\source{
\url{https://www.kaggle.com/c/nfl-big-data-bowl-2022/data}
}
\usage{
PLAYERS
}
\description{
A dataset containing the information on players such as name, height, weight,
date of birth etc.
}
\keyword{datasets}
