% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_plots.R
\name{plot_extra_point}
\alias{plot_extra_point}
\title{Extra Points Plots}
\usage{
plot_extra_point(df, plot_type = "distribution")
}
\arguments{
\item{df}{A dataframe containing information on NFL plays}

\item{plot_type}{Plot one of distribution, successful or unsuccesful}
}
\value{
A ggplot figure containing extra point plots
}
\description{
Extra Points Plots
}
