% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_plots.R
\name{plot_hangTime}
\alias{plot_hangTime}
\title{Hang Time Plots}
\usage{
plot_hangTime(df, playType)
}
\arguments{
\item{df}{A dataframe containing information on NFL plays}

\item{playType}{One of the NFL playtypes Extra Point, Field Goal, Kickoff or
Punt.}
}
\value{
A ggplot figure containing histogram of hangTime
}
\description{
Hang Time Plots
}
