% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_plots.R
\name{plot_kick_result}
\alias{plot_kick_result}
\title{Kick Length Distribution By Play Result Plots}
\usage{
plot_kick_result(df, playType)
}
\arguments{
\item{df}{A dataframe containing information on NFL plays}

\item{playType}{One of the NFL playtypes Extra Point, Field Goal, Kickoff or
Punt.}
}
\value{
A ggplot figure containing `specialTeamsResult` vs. `kickLength`
  comparison
}
\description{
Kick Length Distribution By Play Result Plots
}
