% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_plots.R
\name{plot_results}
\alias{plot_results}
\title{Yards Gained By Play Result Plots}
\usage{
plot_results(df, playType)
}
\arguments{
\item{df}{A dataframe containing information on NFL plays}

\item{playType}{One of the NFL playtypes Extra Point, Field Goal, Kickoff or
Punt.}
}
\value{
A ggplot figure containing specialTeamsResult vs playResult
  comparison
}
\description{
Yards Gained By Play Result Plots
}
