% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_plots.R
\name{plot_team_yardage}
\alias{plot_team_yardage}
\title{Return Yards By Opponent Teams Plots}
\usage{
plot_team_yardage(df, playType)
}
\arguments{
\item{df}{A dataframe containing information on NFL plays}

\item{playType}{One of the NFL playtypes Extra Point, Field Goal, Kickoff or
Punt.}
}
\value{
A ggplot figure containing kickReturnYardage v/s possessionTeam
  comparison
}
\description{
Return Yards By Opponent Teams Plots
}
