% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_plots.R
\name{plot_top_punters}
\alias{plot_top_punters}
\title{Correct Snaps By Punter Plot}
\usage{
plot_top_punters(df, playType)
}
\arguments{
\item{df}{A dataframe containing information on NFL plays}

\item{playType}{One of the NFL playtypes Extra Point, Field Goal, Kickoff or
Punt.}
}
\value{
A ggplot figure containing top Punter players
}
\description{
Correct Snaps By Punter Plot
}
