% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TRACKING}
\alias{TRACKING}
\title{Contains player tracking data from season 2020.}
\format{
A data frame with 41561 rows and 19 variables: \describe{
  \item{time}{Time stamp of play (time, yyyy-mm-dd, hh:mm:ss)}
  \item{x}{Player position along the long axis of the field, 0 - 120 yards.}
  \item{y}{Player position along the short axis of the field, 0 - 53.3
  yards.} \item{s}{Speed in yards/second} \item{a}{Speed in yards/second^2}
  \item{dis}{Distance traveled from prior time point, in yards}
  \item{o}{Player orientation (deg), 0 - 360 degrees} \item{dir}{Angle of
  player motion (deg), 0 - 360 degrees} \item{event}{Tagged play details,
  including moment of ball snap, pass release, pass catch, tackle, etc}
  \item{nflId}{Player identification number, unique across players}
  \item{displayName}{Player name} \item{jerseyNumber}{Jersey number of
  player} \item{position}{Player position group} \item{team}{Team (away or
  home) of corresponding player} \item{frameId}{Frame identifier for each
  play, starting at 1} \item{gameId}{Game identifier, unique}
  \item{playId}{Play identifier, not unique across games}
  \item{playDirection}{Direction that the offense is moving} }
}
\source{
\url{https://www.kaggle.com/c/nfl-big-data-bowl-2022/data}
}
\usage{
TRACKING
}
\description{
A dataset containing the tracking location of different players across
various games
}
\keyword{datasets}
