% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_plots.R
\name{plot_down_kick}
\alias{plot_down_kick}
\title{Title}
\usage{
plot_down_kick(df, playType)
}
\arguments{
\item{df}{A dataframe containing information on NFL plays}

\item{playType}{One of the NFL playtypes Extra Point, Field Goal, Kickoff or Punt.}
}
\value{
A ggplot figure containing kick length v/s down comparison
}
\description{
Title
}
